requirements.txt
# Core ML dependencies - pinned versions for PaddleOCR-VL compatibility
transformers==4.56.2
trl==0.22.2
unsloth==2024.12.10
torch>=2.0.0
xformers
peft>=0.10.0
datasets==4.3.0
bitsandbytes
accelerate
triton
cut_cross_entropy
unsloth_zoo
einops

# Data processing
pillow>=10.0.0
pandas>=2.0.0
sentencepiece
protobuf
huggingface_hub>=0.34.0
hf_transfer

# Monitoring & utilities
wandb>=0.16.0
tenacity>=8.0.0
editdistance>=0.8.1
